+++ server/db/migrations/20200507054829_expense.js
@@ -7,7 +7,7 @@ exports.up = (knex) =>
     t.string('expense_name')
     t.string('categories')
     t.string('frequency').defaultTo('Weekly')
-    t.integer('expense_amount')
+    t.decimal('expense_amount')
     t.boolean('active').defaultTo(true)
   })
